+++ src/styles.scss
@@ -7,11 +7,24 @@ header {
   justify-content: center;
 }
 
+main {
+  padding-top: 5%;
+  display: flex;
+ flex-direction: column;
+ align-items: center;
+}
+
+main p {
+  font-size: 150%;
+  font-weight: bolder;
+  font-family: monospace;
+}
+
 #content {
   width: 100vw;
   height: 100vh;
   color: white;
-  background: linear-gradient(rgba(0,0,0,40%),rgba(0,0,0,40%)), url(./img/restaurant.jpeg);
+  background: linear-gradient(rgba(0,0,0,50%),rgba(0,0,0,50%)), url(./img/restaurant.jpeg);
   background-size: cover;
 }
 
@@ -22,4 +35,8 @@ button {
   border: none;
   padding: 8px;
   margin: 5px;
+}
+
+.logo {
+  width: 300px;
 }
