+++ webpack.config.js
@@ -2,7 +2,7 @@
 
  module.exports = {
   mode: 'development',
-  // devtool: 'inline-source-map',
+  devtool: 'inline-source-map',
    devServer: {
      contentBase: './dist',
    },
